requirements.txt
langchain==0.1.16
langchain-community==0.0.38
langchain-openai==0.1.4
streamlit==1.33.0
python-dotenv==1.0.1
loguru==0.7.2
watchdog==4.0.0
pypdf==4.3.1
chromadb==0.4.24
